+++ cache-manager/src/CacheTypeEnumerate/index.js
@@ -1,5 +1,0 @@
-export default {
-  MEMORY: 0,
-  LOCAL: 1,
-  SESSION: 2
-};