+++ cache-manager/src/cache-type-enumerate/index.js
@@ -1,0 +1,18 @@
+/**
+ * Defines cache type enumerate
+ *
+ * @summary cache-type-enumerate
+ * @version 1.0.0
+ * @author Arian Khosravi <arian.khosravi@aofl.com>
+ */
+
+ const cacheTypeEnumerate = {
+  MEMORY: 0,
+  LOCAL: 1,
+  SESSION: 2
+};
+
+
+export {
+  cacheTypeEnumerate
+};